Validaciones ,Metodo Area y Perimetro
Se haces los metodos segun logica de triangulo

+++ figurasHerencia/src/figurasherencia/Cuadrado.java
@@ -11,10 +11,8 @@ package figurasherencia;
  */
 public class Cuadrado  extends Cuadrilatero{
 
-    public Cuadrado(int coordenada1X, int coordenada1Y, int coordenada2X, int coordenada2Y, int coordenada3X, int coordenada3Y, int coordenada4X, int coordenada4Y) {
+    public Cuadrado(double coordenada1X, double coordenada1Y, double coordenada2X, double coordenada2Y, double coordenada3X, double coordenada3Y, double coordenada4X, double coordenada4Y) {
         super(coordenada1X, coordenada1Y, coordenada2X, coordenada2Y, coordenada3X, coordenada3Y, coordenada4X, coordenada4Y);
     }
-    
-    
     
 }
